# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  auth.LoginRequest:
    properties:
      email:
        example: tanaka@example.com
        type: string
      password:
        example: password123
        maxLength: 50
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      token:
        example: "1234567890"
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      description:
        example: 野球が好きです！
        type: string
      name:
        example: tanaka taro
        maxLength: 100
        minLength: 1
        type: string
      password:
        example: password123
        maxLength: 50
        minLength: 6
        type: string
      profileImage:
        example: https://ik.imagekit.io/your_imagekit_id/image.jpg
        type: string
      token:
        example: "1234567890"
        type: string
      username:
        example: user123
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    - password
    - token
    - username
    type: object
  auth.ResetPassRequest:
    properties:
      afterPassword:
        example: password456
        maxLength: 50
        minLength: 6
        type: string
      token:
        example: "1234567890"
        type: string
    required:
    - afterPassword
    - token
    type: object
  auth.UpdatePassRequestBody:
    properties:
      afterPassword:
        example: password456
        maxLength: 50
        minLength: 6
        type: string
      beforePassword:
        example: password123
        maxLength: 50
        minLength: 6
        type: string
    required:
    - afterPassword
    - beforePassword
    type: object
...
  utils.ApiResponse-auth_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/auth.LoginResponse'
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  utils.ErrorBasicResponse:
    properties:
      message:
        example: エラーメッセージ
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.SuccessBasicResponse:
    properties:
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: このapiは、ビジターゴーのAPIで、ユーザー、スタジアム、遠征記録、などについて扱います。
  title: ビジターゴーAPI
  version: "1.0"
paths:
  /api/admin/auth/emailVerification:
    get:
      description: リクエストからメールアドレス取得後、tokenTypeに応じてチェックし、メールアドレス宛にtokenを含めた画面URLをメールで送信
      parameters:
      - description: メールアドレス
        in: query
        name: email
        required: true
        type: string
      - description: トークンタイプ register or reset
        in: query
        name: tokenType
        required: true
        type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: メールアドレスの本人確認
      tags:
      - auth
  /api/admin/auth/login:
    post:
      description: メールアドレスとパスワードが合致したら、jwtトークンをクライアントに返却
      parameters:
      - description: ログイン情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_LoginResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: ログイン
      tags:
      - auth
  /api/admin/auth/register:
    post:
      description: メールアドレス確認後にリクエスト内容をユーザーテーブルに保存
      parameters:
      - description: ユーザー情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: ユーザー登録
      tags:
      - auth
  /api/admin/auth/resetPass:
    put:
      description: メール内リンクで本人確認後、トークンと新しいパスワードをリクエストで取得し、パスワードを更新する
      parameters:
      - description: tokenと新しいパスワード
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPassRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: ログアウト状態からパスワードを変更
      tags:
      - auth
  /api/admin/auth/updatePass/{userId}:
    put:
      description: 現在のパスワードと新しいパスワードをリクエストで取得し、現在のパスワードが合致したら、新しいパスワードに更新する
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: integer
      - description: メールアドレス
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.UpdatePassRequestBody'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: ログイン状態からパスワードを変更
      tags:
      - auth
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"